Version 1.2 (#2)
* - New translate function which have partitions and delay between it to avoid timeout
- Add async wrap

* Major change:
- Now use deep_translator as base library, deprecate googletrans, support google, microsoft, and mymemory for now
- Add google, microsoft and mymemory support languages json
- Add skip sleep time if sleep time is 0
- Try catch and skip the error

* - Export all constant property (key, sleep time, ...) to config.json and add class to modelize it

* - Add debug mode config
- Add translate microsoft with support multikeys config

* Implement translate with multi translator

* Fix typo

* Implement modify existing localized file on ios and android (+1 squashed commit)
Squashed commits:
[f1dc9a7] Implement modify existing localized file on ios and android

* Update doc and split modify localized files config option

---------

+++ LocalizedGenerator.py
@@ -1,23 +1,23 @@
+from functools import partial, wraps
 from typing import List
-from TranslateContent import LANGUAGES, TranslateContent
+from GeneratorConfig import GeneratorConfig
+from TranslateContent import TranslateContent
+from GGTransLanguage import GGTRANS_LANGUAGE
 import os.path
 import time
 import asyncio
+import io
 
 contents: List[TranslateContent] = []
-iosOutputName = 'output_ios.txt'
-androidOutputName = 'output_android.txt'
-translateLanguages: List[str]= [
-    'id', 'th', 'hi', 'vi', 'es', 'tl',
-    'bn', 'ur', 'zh-tw', 'ja'
-]
- 
 totalContentToTranslate = 0
 translatedContentCount = 0
+config = GeneratorConfig.shared()
     
 # ========= MAIN =========
 async def main():
-    myfile = open('input.txt')
+    config.loadFromJson()
+    
+    myfile = open(config.inputFile)
     myfileContent = ''
     for line in myfile:
         if not line.strip():
@@ -28,28 +28,55 @@ async def main():
     print(myfileContent)
     print()
 
-    await beginTranslate()
-    writeFile('ios')
-    writeFile('android')
-
-async def beginTranslate():
+    await beginTranslateV2()
+    
+    if config.isModifyingExistingLocalizedFilesIOS:
+        modifyExistingFiles('ios')
+    else:
+        writeFile('ios')
+        
+    if config.isModifyingExistingLocalizedFilesAndroid:
+        modifyExistingFiles('android')
+    else:
+        writeFile('android')
+    
+async def beginTranslateV2():
     global totalContentToTranslate
     global translatedContentCount
     
-    print(f'generating localized')
-    totalContentToTranslate = len(list(filter(lambda x: x.isComment == False,contents))) * len(translateLanguages)
+    print(f'generating localized using {config.translatorUsed()}')
+    totalContentToTranslate = len(list(filter(lambda x: x.isComment == False,contents))) * len(GeneratorConfig.shared().outputLanguages)
     translatedContentCount = 0
-    
-    asyncWorks = []
     translatingIndex = 0
-    for language in translateLanguages:
+    
+    asyncPartition = []
+    partitionCount = 0
+    for language in config.outputLanguages:
         translatingIndex = 0
         for (index, item) in enumerate(contents):
             if (item.isComment != True):
                 translatingIndex += 1
-            asyncWorks.append(beginTranslateItem(item=item, language=language))
-    await asyncio.gather(*asyncWorks)
-    print()
+                partitionCount += 1
+            asyncPartition.append(beginTranslateItem(item=item, language=language))
+            
+            if (partitionCount == config.partitionSize):
+                print(f'begin translate current partition')
+                await asyncio.gather(*asyncPartition)
+                print()
+                asyncPartition = []
+                partitionCount = 0
+                
+                if (GeneratorConfig.shared().sleepTime != 0):
+                    print(f'start timeout sleeping in {GeneratorConfig.shared().sleepTime}')
+                    await asyncSleep(GeneratorConfig.shared().sleepTime)
+                
+    if (partitionCount != 0):
+        print(f'begin translate remaining works')
+        await asyncio.gather(*asyncPartition)
+        print()
+        asyncPartition = []
+        partitionCount = 0
+    print(f'done generating localized')
         
 async def beginTranslateItem(item: TranslateContent, language: str):
     global translatedContentCount
@@ -59,7 +86,7 @@ async def beginTranslateItem(item: TranslateContent, language: str):
         print(f"translated {translatedContentCount}/{totalContentToTranslate}", end='\r')
     
 def writeFile(platform: str):
-    filePath = iosOutputName if platform == "ios" else androidOutputName
+    filePath = config.outputIOSFile if platform == "ios" else config.outputAndroidFile
     permission = os.path.exists(filePath) and 'w' or 'x'
     if permission == 'w':
         file = open(filePath, permission)
@@ -68,9 +95,9 @@ def writeFile(platform: str):
     
     print(f'writing file for {platform} localized')
     translatingIndex = 0
-    for language in translateLanguages:
+    for language in config.outputLanguages:
         translatingIndex = 0
-        file.write(f'//{LANGUAGES[language]}\n')
+        file.write(f'//{GGTRANS_LANGUAGE[language]}\n')
         for (index, item) in enumerate(contents):
             if (item.isComment != True):
                 translatingIndex += 1
@@ -78,6 +105,49 @@ def writeFile(platform: str):
             file.write(f'{result}\n')
         file.write('\n')
     file.close()
+    
+def modifyExistingFiles(platform: str):
+    print(f'modifying existing files for {platform} localized')
+    for (index, language) in  enumerate(config.outputLanguages):
+        filePaths: list[str] = config.outputModifyFilesIOS[index].paths
+        if (platform == "android"):
+            filePaths: list[str] = config.outputModifyFilesAndroid[index].paths
+        
+        for filePath in filePaths:
+            if not os.path.exists(filePath):
+                print(f"{filePath} doesn't exist")
+                exit(1)
+            file = open(filePath, "rb+")
+            file.seek(0, io.SEEK_END)
+            if platform == "android":
+                try:  # catch OSError in case of a one line file
+                    file.seek(-len('</resources>'.encode('utf8'))-1, io.SEEK_CUR)
+                except IOError:
+                    pass
+            else:
+                file.write('\n'.encode('utf8'))
+                
+            translatingIndex = 0
+            for (index, item) in enumerate(contents):
+                if (item.isComment != True):
+                    translatingIndex += 1
+                result = item.resultContent(language=language, platform=platform)
+                file.write(f'{result}\n'.encode('utf8'))
+                
+            if platform == "android":
+                file.write('</resources>'.encode('utf8'))
+            file.close()
+        
+def async_wrap(func):
+    @wraps(func)
+    async def run(*args, loop=None, executor=None, **kwargs):
+        if loop is None:
+            loop = asyncio.get_event_loop()
+        pfunc = partial(func, *args, **kwargs)
+        return await loop.run_in_executor(executor, pfunc)
+    return run 
+
+asyncSleep = async_wrap(time.sleep)
         
 if __name__ == '__main__':
     start_time = time.perf_counter()
